requirements.txt
google-auth>=2.29,<3.0
requests>=2.31,<3.0
